backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
flask-restx>=1.3.0
flask-migrate>=4.0.0